% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-table-calc.R
\name{life_table}
\alias{life_table}
\title{Function for estimating period life tables}
\usage{
life_table(x, nMx)
}
\arguments{
\item{x}{a numeric vector of ages}

\item{nMx}{a numeric vector of crude death rates at age x}

\item{n}{width of the intervals}
}
\value{
A period life table given ages \code{x}     and death rates \code{nMx}.
}
\description{
Function for estimating period life tables
}
\examples{
x <- c(0,1,5,10,15,20,25,35,45,55,65,75,85)
females <- read.table(file=
"http://www.demog.berkeley.edu
/~eddieh/AppliedDemographyToolbox
/StanfordCourseLifeTable,
/StanfordCourseMortalityData.csv"
header=TRUE,sep=",")
nMx <- females$Death.Count / females$Population
lifetable(x,nMx)

}
